Updated code. Finished HW05
