+++ app.js
@@ -3,8 +3,7 @@ const logger = require('morgan')
 const cors = require('cors')
 require('dotenv').config();
 
-const authRouter = require('./routes/api/auth');
-const contactsRouter = require('./routes/api/contacts');
+const { contactsRouter, authRouter } = require('./routes');
 
 const app = express()
