+++ controllers/index.js
@@ -1,0 +1,7 @@
+const ctrlContacts = require('./contacts');
+const ctrlUsers = require('./auth');
+
+module.exports = {
+    ctrlContacts,
+    ctrlUsers
+};
