+++ routes/api/auth.js
@@ -1,6 +1,13 @@
 const express = require('express');
 
-const controller = require('../../controllers/auth');
+const { ctrlUsers: {
+    register,
+    login,
+    logout,
+    getCurrent,
+    updateSubscription,
+    updateAvatar,
+}, } = require('../../controllers');
 const { validateBody, validationRegistration, authenticate, upload } = require('../../middlewares');
 const { schemas } = require('../../models/user');
 
@@ -8,16 +15,16 @@ const router = express.Router();
 
 
 // signup
-router.post("/register", validationRegistration(schemas.registerSchema), controller.register);
+router.post("/register", validationRegistration(schemas.registerSchema), register);
 // signin
-router.post("/login", validationRegistration(schemas.loginSchema), controller.login);
+router.post("/login", validationRegistration(schemas.loginSchema), login);
 
-router.post("/logout", authenticate, controller.logout);
+router.post("/logout", authenticate, logout);
 
-router.get("/current", authenticate, controller.getCurrent);
+router.get("/current", authenticate, getCurrent);
 
-router.patch('/', authenticate, validateBody(schemas.updateSubscriptionSchema), controller.updateSubscription);
+router.patch('/', authenticate, validateBody(schemas.updateSubscriptionSchema), updateSubscription);
 
-router.patch('/avatars', authenticate, upload.single('avatar'), controller.updateAvatar);
+router.patch('/avatars', authenticate, upload.single('avatar'), updateAvatar);
 
 module.exports = router;
