+++ routes/index.js
@@ -1,0 +1,7 @@
+const contactsRouter = require('./api/contacts');
+const authRouter = require('./api/auth');
+
+module.exports = {
+    contactsRouter,
+    authRouter,
+};
